feat(player-turn): player turn

See merge request polytechnique-montr-al/log2990/20243/equipe-101/LOG2990-101!44

+++ server/app/services/game/game.service.ts
@@ -1,30 +1,15 @@
-import { Timer } from '@app/classes/timer/timer';
-import { FIGHT_TIME, STARTING_TIME } from '@app/constants';
+import { FELLING_PROBABILITY, MOVEMENT_TIME, SINGLE_PLAYER, STARTING_TIME, TileCost, TileType, TURN_TIME } from '@app/constants';
 import { RoomService } from '@app/services/room/room.service';
-import { Avatar, Player, Status } from '@common/player';
-import { Room } from '@common/room';
+import { Avatar, Player, Position, Status } from '@common/player';
+import { GameStatus, Room } from '@common/room';
 import { Injectable } from '@nestjs/common';
 import { Server, Socket } from 'socket.io';
 
 @Injectable()
 export class GameService {
-    turnTimer = new Timer();
-    fightTimer = new Timer();
-
+    isMoving: boolean = false;
+    isTurnSkipped: boolean = false;
     constructor(private roomService: RoomService) {}
-
-    toggleLockRoom(roomId: string, isLocked: boolean) {
-        const game = this.getGame(roomId);
-        game.isLocked = isLocked;
-    }
-
-    getGame(roomId) {
-        return this.roomService.rooms.get(roomId);
-    }
-
-    isCodeFormatValid(roomCode: string): boolean {
-        return /^[0-9]{4}$/.test(roomCode);
-    }
 
     connectPlayerToGame(roomId: string) {
         const game = this.getGame(roomId);
@@ -51,40 +36,31 @@ export class GameService {
         takenAvatar.isTaken = true;
     }
 
-    setUniquePlayerName(player: Player, socket: Socket) {
-        const playerName = this.generateUniquePlayerName(player.name, socket);
-        player.name = playerName;
-        socket.data.username = player.name;
+    getActivePlayer(room: Room): Player {
+        return room.listPlayers.find((player) => player.isActive === true);
     }
 
-    isActivePlayer(socket: Socket) {
-        const room = this.roomService.getRoom(socket);
-        const currentPlayer = room.listPlayers.find((player) => player.id === socket.id);
-        return currentPlayer.isActive;
+    getGame(roomId) {
+        return this.roomService.rooms.get(roomId);
+    }
+
+    getPlayerById(room: Room, socket: Socket) {
+        return room.listPlayers.find((player) => player.id === socket.id);
     }
 
     leavePlayerFromGame(roomId: string, socket: Socket, server: Server) {
         const isAdmin = this.roomService.isPlayerAdmin(socket);
         const room = this.roomService.getRoom(socket);
         socket.emit('leftRoom', isAdmin);
-        if (isAdmin) {
+        if (isAdmin && room.gameStatus === GameStatus.Lobby) {
             this.roomService.deleteRoom(roomId, socket);
+        } else if (room.gameStatus === GameStatus.Started) {
+            this.playerDisconnected(room, socket, server);
+            socket.to(roomId).emit('disconnectedPlayer', room.listPlayers);
         } else {
             this.removePlayerFromRoom(roomId, socket, server);
             socket.to(roomId).emit('updatedPlayer', room);
         }
-    }
-
-    removePlayerFromRoom(roomId: string, socket: Socket, server: Server) {
-        const room = this.roomService.rooms.get(roomId);
-        room.listPlayers = room.listPlayers.filter((player) => player.id !== socket.id);
-        this.freeUpAvatar(room, socket);
-        this.updateAvatarsForAllClients(server, roomId);
-        this.roomService.leaveRoom(roomId, socket);
-    }
-
-    getAvatarByName(room: Room, avatar: Avatar) {
-        return room.availableAvatars.find((av) => av.name === avatar.name);
     }
 
     selectedAvatar(room: Room, avatar: Avatar, socket: Socket, server: Server) {
@@ -97,7 +73,106 @@ export class GameService {
         }
     }
 
-    freeUpAvatar(room: Room, socket: Socket) {
+    stopGameTimers(room: Room) {
+        this.roomService.getFightTimer(room.roomId).stopTimer();
+        this.roomService.getTurnTimer(room.roomId).stopTimer();
+    }
+
+    removePlayerFromRoom(roomId: string, socket: Socket, server: Server) {
+        const room = this.roomService.rooms.get(roomId);
+        room.listPlayers = room.listPlayers.filter((player) => player.id !== socket.id);
+        this.freeUpAvatar(room, socket);
+        this.updateAvatarsForAllClients(server, roomId);
+        this.roomService.leaveRoom(roomId, socket);
+    }
+
+    toggleLockRoom(roomId: string, isLocked: boolean) {
+        const game = this.getGame(roomId);
+        game.isLocked = isLocked;
+    }
+
+    onStartGame(room: Room) {
+        room.gameStatus = GameStatus.Started;
+        this.sortPlayersBySpeed(room);
+        room.listPlayers[0].isActive = true;
+    }
+
+    onStartTurn(client: Socket, server: Server) {
+        const room = this.roomService.getRoom(client);
+        const activePlayer = this.getActivePlayer(room);
+        server.to(room.roomId).emit('otherPlayerTurn', activePlayer.name);
+        this.roomService.getTurnTimer(room.roomId).startTimer(STARTING_TIME, (timeRemaining) => {
+            server.to(activePlayer.id).emit('beforeStartTurnTimer', timeRemaining);
+            if (timeRemaining <= 0) {
+                this.playerTurnTimer(client, server);
+            }
+        });
+    }
+
+    onTurnEnded(client: Socket, server: Server) {
+        const room = this.roomService.getRoom(client);
+        if (!this.isMoving) {
+            this.updateActivePlayer(client);
+            const activePlayer = this.getActivePlayer(room);
+            server.to(room.roomId).emit('isActive', activePlayer.id);
+            server.to(room.roomId).emit('turnEnded', room.listPlayers);
+        } else {
+            this.isTurnSkipped = true;
+        }
+    }
+
+    async processNavigation(room: Room, server: Server, path: Position[], client: Socket) {
+        const player = this.getActivePlayer(room);
+        for (const tile of path) {
+            this.isMoving = true;
+            player.position = tile;
+            if (this.isMoving) {
+                await this.delay(MOVEMENT_TIME);
+            }
+            server.to(room.roomId).emit('playerNavigation', tile);
+            if (room.gameMap.tiles[tile.x][tile.y] === TileType.Ice && !this.checkFell()) {
+                this.stopGameTimers(room);
+                client.emit('playerFell');
+                break;
+            }
+            if (room.gameMap.tiles[tile.x][tile.y] !== TileType.Ice) {
+                player.attributes.movementPointsLeft -= this.getCost(room.gameMap.tiles[tile.x][tile.y]);
+            }
+        }
+        this.getActivePlayer(room).attributes.movementPointsLeft = this.getActivePlayer(room).attributes.speed;
+        server.to(room.roomId).emit('endMovement');
+        this.isMoving = false;
+        if (this.isTurnSkipped) {
+            this.onTurnEnded(client, server);
+            this.isTurnSkipped = false;
+        }
+    }
+
+    async delay(ms: number) {
+        return new Promise((resolve) => setTimeout(resolve, ms));
+    }
+
+    private getCost(tileType: number): number {
+        switch (tileType) {
+            case TileType.Ground:
+                return TileCost.Ground;
+            case TileType.Water:
+                return TileCost.Water;
+            case TileType.Ice:
+                return TileCost.Ice;
+            case TileType.OpenDoor:
+                return TileCost.OpenDoor;
+            default:
+                return Infinity;
+        }
+    }
+
+    private checkFell(): boolean {
+        const randomValue = Math.random();
+        return randomValue > FELLING_PROBABILITY;
+    }
+
+    private freeUpAvatar(room: Room, socket: Socket) {
         if (socket.data.clickedAvatar) {
             const previousAvatar = this.getAvatarByName(room, socket.data.clickedAvatar);
             if (previousAvatar) {
@@ -106,15 +181,68 @@ export class GameService {
         }
     }
 
-    updateAvatarsForAllClients(server: Server, roomId: string) {
-        server.sockets.sockets.forEach((clientSocket: Socket) => {
-            if (clientSocket.rooms.has(roomId)) {
-                this.sendAvatarListToClient(clientSocket);
+    private generateUniquePlayerName(playerName: string, socket: Socket): string {
+        let name = playerName;
+        let suffix = 2;
+
+        while (this.isPlayerNameTaken(name, socket)) {
+            name = `${playerName}-${suffix}`;
+            suffix++;
+        }
+        return name;
+    }
+
+    private getAvatarByName(room: Room, avatar: Avatar) {
+        return room.availableAvatars.find((av) => av.name === avatar.name);
+    }
+
+    private getPlayerConnectedInRoom(room: Room) {
+        return room.listPlayers.filter((player) => player.status !== Status.Disconnected);
+    }
+
+    private isActivePlayer(socket: Socket) {
+        const room = this.roomService.getRoom(socket);
+        const currentPlayer = this.getPlayerById(room, socket);
+        return currentPlayer.isActive;
+    }
+
+    private isLastPlayer(room: Room) {
+        return this.getPlayerConnectedInRoom(room).length === SINGLE_PLAYER;
+    }
+
+    private isCodeFormatValid(roomCode: string): boolean {
+        return /^[0-9]{4}$/.test(roomCode);
+    }
+
+    private isPlayerNameTaken(name: string, socket: Socket) {
+        const playersList = this.roomService.getRoom(socket).listPlayers;
+        return playersList.some((player) => player.name === name);
+    }
+
+    private playerDisconnected(room: Room, socket: Socket, server: Server) {
+        const disconnectedPlayer = this.getPlayerById(room, socket);
+        if (this.isActivePlayer(socket)) {
+            this.onTurnEnded(socket, server);
+        }
+        disconnectedPlayer.status = Status.Disconnected;
+        if (this.isLastPlayer(room)) {
+            server.to(room.roomId).emit('draw');
+        }
+        server.to(room.roomId).emit('playerDisconnected', disconnectedPlayer);
+        this.sortPlayersBySpeed(room);
+    }
+
+    private playerTurnTimer(client: Socket, server: Server) {
+        const room = this.roomService.getRoom(client);
+        this.roomService.getTurnTimer(room.roomId).resetTimer(TURN_TIME, (timeRemaining) => {
+            server.to(room.roomId).emit('startedTurnTimer', timeRemaining);
+            if (timeRemaining <= 0) {
+                this.onTurnEnded(client, server);
             }
         });
     }
 
-    sendAvatarListToClient(socket: Socket) {
+    private sendAvatarListToClient(socket: Socket) {
         const room = this.roomService.getRoom(socket);
         const customizedAvatarsList = room.availableAvatars.map((avatar) => {
             const isSelectedByClient = socket.data.clickedAvatar?.name === avatar.name;
@@ -124,79 +252,13 @@ export class GameService {
                 isSelected: isSelectedByClient,
             };
         });
-
         socket.emit('characterSelected', customizedAvatarsList);
     }
 
-    updateActivePlayer(socket: Socket) {
-        const room = this.roomService.getRoom(socket);
-        const listPlayers = room.listPlayers;
-        const index = listPlayers.findIndex((item) => item.id === socket.id);
-        const nextIndex = (index + 1) % listPlayers.length;
-        listPlayers[index].isActive = false;
-        listPlayers[nextIndex].isActive = true;
-    }
-
-    onStartGame(room: Room) {
-        this.sortPlayersBySpeed(room);
-    }
-
-    onStartTurn(room: Room, server: Server) {
-        const activePlayer = room.listPlayers.find((player) => player.isActive === true);
-        this.turnTimer.startTimer(STARTING_TIME, (timeRemaining) => {
-            server.to(activePlayer.id).emit('beforeStartTurnTimer', timeRemaining);
-            if (timeRemaining === 0) {
-                server.to(activePlayer.id).emit('beforeStartTurnTimerEnd');
-            }
-        });
-    }
-
-    onPlayerTurnStarted(duration: number, client: Socket, server: Server) {
-        const room = this.roomService.getRoom(client);
-        this.turnTimer.startTimer(duration, (timeRemaining) => {
-            server.to(room.roomId).emit('startedTurnTimer', timeRemaining);
-            if (timeRemaining === 0) {
-                this.onTurnEnded(client, server);
-                server.to(room.roomId).emit('turnEnded', room.listPlayers);
-            }
-        });
-    }
-
-    onTurnEnded(client: Socket, server: Server) {
-        const room = this.roomService.getRoom(client);
-        this.updateActivePlayer(client);
-        const activePlayer = room.listPlayers.find((player) => player.isActive === true);
-        server.to(room.roomId).emit('isActive', activePlayer.id);
-    }
-
-    onStartFight(client: Socket, opponent: Player, server: Server) {
-        this.turnTimer.pauseTimer();
-        this.fightTimer.startTimer(FIGHT_TIME, (timeRemaining) => {
-            client.emit('fightTime', timeRemaining);
-            server.to(opponent.id).emit('fightTime', timeRemaining);
-        });
-    }
-
-    onEndFight(server: Server, room: Room) {
-        this.turnTimer.resumeTimer((timeRemaining) => {
-            server.to(room.roomId).emit('startedTurnTimer', timeRemaining);
-        });
-    }
-
-    isPlayerNameTaken(name: string, socket: Socket) {
-        const playersList = this.roomService.getRoom(socket).listPlayers;
-        return playersList.some((player) => player.name === name);
-    }
-
-    generateUniquePlayerName(playerName: string, socket: Socket): string {
-        let name = playerName;
-        let suffix = 2;
-
-        while (this.isPlayerNameTaken(name, socket)) {
-            name = `${playerName}-${suffix}`;
-            suffix++;
-        }
-        return name;
+    private setUniquePlayerName(player: Player, socket: Socket) {
+        const playerName = this.generateUniquePlayerName(player.name, socket);
+        player.name = playerName;
+        socket.data.username = player.name;
     }
 
     private sortPlayersBySpeed(room: Room) {
@@ -208,6 +270,23 @@ export class GameService {
                 ...listPlayers.filter((player) => player.status === Status.Disconnected),
             ];
         }
-        listPlayers[0].isActive = true;
+        room.listPlayers = listPlayers;
+    }
+
+    private updateActivePlayer(socket: Socket) {
+        const room = this.roomService.getRoom(socket);
+        const listPlayers = this.getPlayerConnectedInRoom(room);
+        const index = listPlayers.findIndex((item) => item.id === this.getActivePlayer(room).id);
+        const nextIndex = (index + 1) % listPlayers.length;
+        listPlayers[index].isActive = false;
+        listPlayers[nextIndex].isActive = true;
+    }
+
+    private updateAvatarsForAllClients(server: Server, roomId: string) {
+        server.sockets.sockets.forEach((clientSocket: Socket) => {
+            if (clientSocket.rooms.has(roomId)) {
+                this.sendAvatarListToClient(clientSocket);
+            }
+        });
     }
 }
